% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.utils-package.R
\docType{package}
\name{dc.utils-package}
\alias{dc.utils}
\alias{dc.utils-package}
\title{dc.utils: Data Commons Utility Functions}
\description{
Common Utility Functions Supporting Data Repository Construction and Maintenance
}
\author{
\strong{Maintainer}: Aaron Schroeder \email{dads2busy@gmail.com}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
  \item Kathryn Linehan \email{kjl5t@virginia.edu}
  \item Micah Iserman \email{micah.iserman@gmail.com}
}

}
\keyword{internal}
