% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvrs_get_dataset_files.R
\name{dvrs_get_dataset_files}
\alias{dvrs_get_dataset_files}
\title{Get Dataverse Dataset File(s)}
\usage{
dvrs_get_dataset_files(
  server = Sys.getenv("DATAVERSE_SERVER"),
  key = Sys.getenv("DATAVERSE_KEY"),
  version = NULL,
  doi = NULL
)
}
\arguments{
\item{server}{Sys.getenv("DATAVERSE_SERVER")}

\item{key}{Sys.getenv("DATAVERSE_KEY")}

\item{version}{version of dataset. options are :draft, :latest, :latest-published, or a specific version number}

\item{doi}{dataset doi or vector of dois}
}
\description{
Will download and unzip all dataset files in the current directory.
}
\examples{
\dontrun{dvrs_get_dataset_files(version = ":draft", doi = "doi:XXX")}
}
