% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvrs_metadata_check.R
\name{dvrs_metadata_check}
\alias{dvrs_metadata_check}
\title{Dataverse Metadata Check}
\usage{
dvrs_metadata_check(
  server = Sys.getenv("DATAVERSE_SERVER"),
  key = Sys.getenv("DATAVERSE_KEY"),
  version = NULL,
  doi = NULL
)
}
\arguments{
\item{server}{Sys.getenv("DATAVERSE_SERVER")}

\item{key}{Sys.getenv("DATAVERSE_KEY")}

\item{version}{version of dataset. options are :draft, :latest, :latest-published, or a specific version number}

\item{doi}{dataset doi or vector of dois}
}
\description{
Checks Dataverse metadata against the SDAD standards and website metadata.
}
\examples{
\dontrun{dvrs_metadata_check(version = ":draft", doi = "doi:XXX")}
}
