% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_checksums.R
\name{file_checksums}
\alias{file_checksums}
\title{Generate MD5 file checksums and save as json file}
\usage{
file_checksums(
  dir_path = "data",
  dir_path_contains = "distribution",
  file_extensions = c("csv", "geojson", "parquet", "tif"),
  output_path = "data",
  output_name = "distribution_file_manifest"
)
}
\arguments{
\item{dir_path}{Path to directory of files to process.}

\item{dir_path_contains}{Path to directory of files contains a specific string.}

\item{file_extensions}{Filter for specific file types like csv or tif}

\item{output_path}{Path to output directory for json file}

\item{output_name}{Name of output file.}
}
\description{
Generate MD5 file checksums and save as json file
}
\examples{
file_checksums("data/")
list.files("data", pattern = "*manifest*")
}
