% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_find_by_type.R
\name{file_find_by_type}
\alias{file_find_by_type}
\title{Find By File Type in Directory}
\usage{
file_find_by_type(
  directory,
  type = "shp",
  first_only = FALSE,
  full_path = FALSE
)
}
\arguments{
\item{type}{Type of file file extension without a period, e.g. "zip", "shp".}

\item{first_only}{Only return first found. default FALSE.}

\item{full_path}{Return file name with full path. defualt FASLE.}
}
\description{
Find By File Type in Directory
}
\examples{
file_find_by_type(tmpdir, "shp", first_only = TRUE)
[1] "some_shape_file.shp"
}
