% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_is_zip.R
\name{is_zip}
\alias{is_zip}
\title{Check if file is a zip file.}
\usage{
is_zip(filepath)
}
\arguments{
\item{filepath}{path to file}
}
\description{
Check if file is a zip file.
}
\examples{
is_zip("a_file.zip")
[1] TRUE
}
