% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_lats_lons2geo_areas.R
\name{loc_lat_lon2geo_areas}
\alias{loc_lat_lon2geo_areas}
\title{Get census block, county, state, and market area information based on latitude/longitude input
from the FCC API.}
\usage{
loc_lat_lon2geo_areas(place_id = "VTRC", lat = 38.880807, lon = -77.11577)
}
\arguments{
\item{place_id}{some unique identifier for the lat lon}

\item{lat}{the latitude}

\item{lon}{the longitude}
}
\value{
data.frame
}
\description{
Get census block, county, state, and market area information based on latitude/longitude input
from the FCC API.
}
\examples{
loc_lat_lon2geo_areas("VTRC", lat=38.880807, lon=-77.11577)
}
