% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_lats_lons2geo_areas.R
\name{loc_lats_lons2geo_areas}
\alias{loc_lats_lons2geo_areas}
\title{Get census block, county, state, and market area information based on multiple latitude/longitude inputs.
#' from the FCC API.}
\usage{
loc_lats_lons2geo_areas(
  place_idCol = c("VTRC", "VT-NVC"),
  latCol = c(38.880807, 38.8968325),
  lonCol = c(-77.11577, -77.1894815)
)
}
\arguments{
\item{place_idCol}{vector of unique identifiers}

\item{latCol}{vector of latitudes}

\item{lonCol}{vector of longitudes}
}
\value{
data.frame
}
\description{
Get census block, county, state, and market area information based on multiple latitude/longitude inputs.
#' from the FCC API.
}
\examples{
loc_lats_lons2geo_areas(place_idCol = c("VTRC", "VT-NVC"),
                  latCol = c(38.880807, 38.8968325),
                  lonCol = c(-77.11577, -77.1894815))
}
