% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_move_suffix.R
\name{set_move_suffix}
\alias{set_move_suffix}
\title{Move provided suffixes to separate column (and remove suffix from name column).}
\usage{
set_move_suffix(x, col_name, suffix_list)
}
\arguments{
\item{x}{A data.frame, data.table or tbl with a suffix containing name column.}

\item{col_name}{Name of the name column/field.}

\item{suffix_list}{A list of suffixes to move.}
}
\description{
Move provided suffixes to separate column (and remove suffix from name column).
}
\examples{
names <- data.frame(name = c("Bob, Jr", "Dallas III", "Willy"))
names <- set_move_suffix(names, "name", list("Jr", "III"))
names
   name suffix
1  Bob,     Jr
2  Dallas    III
3  Willy   <NA>
}
