% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_std_col_names.R
\name{set_std_col_names}
\alias{set_std_col_names}
\title{Standardize column names
Make lower case R compliant names that use an underscore rather than a dot
and remove apostrophes. Multiple underscores are reduced to one.}
\usage{
set_std_col_names(
  name_list = c("first.name", "LastName"),
  fix_acronyms = TRUE,
  fix_camel = FALSE
)
}
\arguments{
\item{name_list}{List of column names.}

\item{fix_camel}{Change camel case to underscores before processing.}
}
\description{
Standardize column names
Make lower case R compliant names that use an underscore rather than a dot
and remove apostrophes. Multiple underscores are reduced to one.
}
\examples{
set_std_col_names(c("first.name", "LastName"))
[1] "first_name" "lastname"
set_std_col_names(c("first.name", "LastName"), fix_camel = T)
[1] "first_name" "last_name"
}
