% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_fix_acronyms.R
\name{str_fix_acronyms}
\alias{str_fix_acronyms}
\title{Change acronyms (multiple capital letters) to title case}
\usage{
str_fix_acronyms(str = c("FIPSCode", "NoAcronymHere", "Multi-APNFlag"))
}
\arguments{
\item{str}{List of colnames to convert.}
}
\description{
Change acronyms (multiple capital letters) to title case
}
\examples{
str_fix_acronyms(c("FIPSCode", "NoAcronymHere", "Multi-APNFlag"))
[1] "FipsCode"      "NoAcronymHere" "Multi-ApnFlag"
}
