% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_fix_camel_case.R
\name{str_fix_camel_case}
\alias{str_fix_camel_case}
\title{Convert camel case to underscore separated words}
\usage{
str_fix_camel_case(name_list = c("I'mACamel"))
}
\arguments{
\item{name_list}{List of camel case strings to convert.}
}
\description{
Convert camel case to underscore separated words
}
\examples{
str_fix_camel_case(c("I'mACamel", "NoYouAreNot"))
[1] "I'm_A_Camel"    "No_You_Are_Not"
}
