% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_remove_non_alphanum.R
\name{str_remove_non_alphanum}
\alias{str_remove_non_alphanum}
\title{Remove non-alphanumeric characters from a string.
Optionally remove spaces.}
\usage{
str_remove_non_alphanum(string, keep_spaces = TRUE)
}
\arguments{
\item{string}{A character string.}

\item{keep_spaces}{Don't remove spaces.}
}
\description{
Remove non-alphanumeric characters from a string.
Optionally remove spaces.
}
\examples{
remove_non_alphanum("222-44-6666")
[1] "222446666"
str_remove_non_alphanum("Devante Smith-Pelly", keep_spaces = F)
[1] "DevanteSmithPelly"
}
