% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_new_dataset_setup.R
\name{utl_new_dataset_setup}
\alias{utl_new_dataset_setup}
\title{Set up directories and sample files for creating a new dataset}
\usage{
utl_new_dataset_setup(
 
    dataset_info_yml_file_path = "src/01_data/00_dataset_yaml_files/fairfax_county_zip_codes_geo.yml",
  code_files = "R",
  overwrite = FALSE
)
}
\arguments{
\item{dataset_info_yml_file_path}{Yaml settings file for new dataset}

\item{code_files}{Code file type preference. Currently only .R}

\item{overwrite}{Overwrite existing directories and sample files}
}
\description{
Set up directories and sample files for creating a new dataset
}
\examples{
new_dataset_setup(dataset_info_yml_file_path = "src/01_data/00_dataset_yaml_files/us_census_counties_geo_2010.yml", code_files = "R", overwrite = TRUE)
}
