% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_ssn.R
\name{valid_ssn}
\alias{valid_ssn}
\title{Check if Social Security Number is Valid.}
\usage{
valid_ssn(ssn)
}
\arguments{
\item{ssn}{Social Security Number to check.}
}
\description{
Check if Social Security Number is Valid.
}
\examples{
valid_ssn("123-45-6789")
[1] FALSE
}
