% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_ssn_add_col.R
\name{valid_ssn_add_col}
\alias{valid_ssn_add_col}
\title{Add a valid_ssn column to a data.frame or data.table.}
\usage{
valid_ssn_add_col(df, ssn_field)
}
\arguments{
\item{df}{data.frame or data.table with an ssn column.}

\item{ssn_field}{Name of the ssn column/field.}
}
\description{
Add a valid_ssn column to a data.frame or data.table.
}
\examples{
ssn_df <- data.frame(ssn = c("123-45-6789", "368-96-8955", "999998888", "287-65-4321"))
ssn_df <- valid_ssn_add_col(ssn_df, "ssn")
ssn_df
ssn ssn_valid
1 123-45-6789    FALSE
2 367-94-8940     TRUE
3   999998888    FALSE
4 287-65-4321     TRUE
}
