% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_usps_address.R
\name{valid_usps_address}
\alias{valid_usps_address}
\title{Validate address with U.S. Postal Service (USPS) Web Tools}
\usage{
valid_usps_address(
  address_1 = "",
  address_2 = "1100 Wilson Blvd",
  city = "Arlington",
  state = "VA",
  zip5 = "",
  zip4 = "",
  full_info = TRUE,
  usps_userid = Sys.getenv("USPS_WEBAPI_USERNAME")
)
}
\arguments{
\item{address_1}{Optional. Secondary Delivery Address. May contain secondary unit designator, such as APT or SUITE.}

\item{address_2}{Optional. Primary Delivery Address.}

\item{city}{Optional. City name of the destination address. Maximum characters allowed: 15.}

\item{state}{Optional. Two-character state code of the destination address. Maximum characters allowed: 2.}

\item{zip5}{Optional. Destination 5-digit ZIP Code.StringMust be 5-digits. Numeric values (0-9) only.}

\item{zip4}{Optional. Destination 4-digit +ZIP Code. Numeric values (0-9) only.}

\item{full_info}{Required. Logical value used to flag return of all response fields.}
}
\value{
list
}
\description{
Validate address with U.S. Postal Service (USPS) Web Tools
}
