% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_dir_exists.R
\name{verify_dir_exists}
\alias{verify_dir_exists}
\title{Verify directory exists (if not, create)}
\usage{
verify_dir_exists(full_path, recursive = TRUE)
}
\arguments{
\item{full_path}{Path to directory.}
}
\description{
Verify directory exists (if not, create)
}
\examples{
verify_dir_exists("~/missingfolder")
Creates folder named "missingfolder" in your home directory
}
