% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_is_blank.R
\name{verify_is_blank}
\alias{verify_is_blank}
\title{Checks if provided variable is null, na, or of length 0}
\usage{
verify_is_blank(x)
}
\arguments{
\item{x}{Variable provided.}
}
\description{
Checks if provided variable is null, na, or of length 0
}
\examples{
varify_is_blank("")
[1] TRUE
}
