% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_get_links.R
\name{web_get_links}
\alias{web_get_links}
\title{Extract link texts and urls from a web page}
\usage{
web_get_links(url)
}
\arguments{
\item{url}{character an url}
}
\value{
a data frame of link text and urls
}
\description{
Extract link texts and urls from a web page
}
\examples{
\dontrun{
web_get_links("http://localhost/")
glinks <- get_webpage_links("http://google.com/")
}
}
