% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.utils-package.R
\docType{package}
\name{dc.utils-package}
\alias{dc.utils}
\alias{dc.utils-package}
\title{dc.utils: Dataverse Standardization Checks}
\description{
Allows user to test their data on the UVA Dataverse for SDAD standards.
}
\author{
\strong{Maintainer}: Kathryn Linehan \email{kjl5t@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
